refactor(sdk): import concrete store modules instead of the stores barrel
The tsup DTS pass emitted 12 Rollup warnings about store exports being
re-exported through src/stores/index.ts while the barrel and the store
modules were mutually dependent. The cycle was type-only (a value-import
graph walk found none), so nothing was broken at runtime, but the warnings
were noise on every SDK build.

Internal SDK modules now import the concrete store file (../stores/chatStore)
rather than the ../stores barrel, which is the consumer entry point. The
barrel also re-exports the event types from ../core/types/events instead of
the ../core/types barrel, removing its last edge into core/. Cycles through
stores/index.ts: 778 -> 0.

An ESLint no-restricted-imports rule keeps it that way; src/index.ts and
test files are exempt.

Two test files mocked the barrel with vi.mock('../stores'), which silently
stopped intercepting once the source imported concrete modules; they now
mock the concrete modules.
