+++ packages/fluux-sdk/src/core/XMPPClient.ts
@@ -26,7 +26,7 @@ import type { ConnectionActor, ConnectionStateValue } from './connectionMachine'
 import { ensureCryptoRandomUUID } from './polyfill'
 import { createStoreBindings } from '../bindings/storeBindings'
 import { setupStoreSideEffects } from './sideEffects'
-import { defaultStores, type SDKStores } from '../stores'
+import { defaultStores, type SDKStores } from '../stores/sdkStores'
 import { detectPlatform } from './platform'
 import { isDeadSocketError } from './modules/connectionUtils'
 import { getBareJid, getDomain } from './jid'
