+++ packages/fluux-sdk/src/core/mdsSideEffects.ts
@@ -25,7 +25,9 @@
 import type { XMPPClient } from './XMPPClient'
 import type { DisplayedMarker } from './modules/Mds'
 import type { SideEffectsOptions } from './chatSideEffects'
-import { chatStore, connectionStore, roomStore } from '../stores'
+import { chatStore } from '../stores/chatStore'
+import { connectionStore } from '../stores/connectionStore'
+import { roomStore } from '../stores/roomStore'
 import { createKeyedCoalescer } from '../utils/keyedCoalescer'
 import { getBareJid } from './jid'
 import { logInfo } from './logger'
