+++ packages/fluux-sdk/src/hooks/useRoomActions.ts
@@ -1,5 +1,5 @@
 import { useCallback, useMemo } from 'react'
-import { roomStore } from '../stores'
+import { roomStore } from '../stores/roomStore'
 import { useXMPPContext } from '../provider'
 import type {
   MentionReference,
